MSP-43 Moved all outfit endpoint to UserController.java

+++ src/main/java/de/gaknr/mspbackend/outfit/OutfitService.java
@@ -1,5 +1,7 @@
 package de.gaknr.mspbackend.outfit;
 
+import de.gaknr.mspbackend.user.UserEntity;
+import de.gaknr.mspbackend.user.UserService;
 import lombok.RequiredArgsConstructor;
 
 import org.bson.types.ObjectId;
@@ -14,13 +16,21 @@ import java.util.Optional;
 public class OutfitService {
 
     private final OutfitRepository repository;
+    private final UserService userService;
 
-    public void save(OutfitEntity outfitEntity) {
-        repository.save(outfitEntity);
+    public void save(OutfitEntity outfitEntity, String userId) {
+        UserEntity entity = userService.getById(userId);
+        entity.getOutfits().add(repository.save(outfitEntity).getId());
+        userService.update(entity, userId);
     }
 
-    public void deleteById(ObjectId id) {
-        repository.deleteById(id);
+    public void deleteById(ObjectId id, String userId) {
+        UserEntity entity = userService.getById(userId);
+        if(entity.getOutfits().contains(id)) {
+            entity.getOutfits().remove(id);
+            userService.update(entity, userId);
+            repository.deleteById(id);
+        }
     }
 
     public OutfitEntity getById(ObjectId id) {
@@ -28,8 +38,13 @@ public class OutfitService {
         return optional.orElse(null);
     }
 
-    public List<OutfitEntity> getAll() {
-        return repository.findAll();
+    public void update(OutfitEntity outfitEntity, ObjectId id) {
+        Optional<OutfitEntity> optional = repository.findById(id);
+        if (optional.isPresent()) {
+            optional.get().setPieces(outfitEntity.getPieces());
+            optional.get().setFavorite(outfitEntity.isFavorite());
+            repository.save(optional.get());
+        }
     }
 
 }
